game_loop and game_update
