+++ src/main.cpp
@@ -132,6 +132,7 @@
 #include "net_input_lag.h"
 #include "moonphase.h"
 #include "frontmenu_ingame_map.h"
+#include "room_library.h"
 #include <cstdint>
 
 #ifdef FUNCTESTING
@@ -144,7 +145,7 @@
 #define strcasecmp _stricmp
 #endif
 
-short do_draw;
+
 short default_loc_player = 0;
 struct StartupParameters start_params;
 char autostart_multiplayer_campaign[80] = "";
@@ -204,11 +205,10 @@ extern "C" {
 TbBool force_player_num = false;
 
 /******************************************************************************/
-extern void faststartup_network_game(CoroutineLoop *context);
+
 extern void faststartup_saved_packet_game(void);
-extern TngUpdateRet damage_creatures_with_physical_force(struct Thing *thing, ModTngFilterParam param);
-extern CoroutineLoopState set_not_has_quit(CoroutineLoop *context);
-extern void startup_network_game(CoroutineLoop *context, TbBool local);
+
+
 /******************************************************************************/
 
 TbClockMSec timerstarttime = 0;
@@ -221,12 +221,11 @@ TbBool TimerFreeze = false;
 int32_t fps_limit_current = 0;
 int32_t fps_limit_main = 0; // -1 if auto
 int32_t fps_limit_secondary = 0;
-static long double process_frame_time = 0;
-static long double time_since_last_draw = 0;
-static long double average_frame_draw_time = 1;
-static long double multiplayer_clock_adjust = 1;
+long double process_frame_time = 0;
+long double time_since_last_draw = 0;
+long double multiplayer_clock_adjust = 1;
 long double host_packet_received = 1;
-float interpolate_time = 0;
+
 /******************************************************************************/
 
 
@@ -237,30 +236,7 @@ void setup_stuff(void)
     init_alpha_table();
 }
 
-TbBool should_use_delta_time_on_menu()
-{
-    switch (frontend_menu_state) {
-        case FeSt_MAIN_MENU:
-        case FeSt_FELOAD_GAME:
-        case FeSt_NET_SERVICE: /**< Network service selection, where player can select Serial/Modem/IPX/TCP IP/1 player. */
-        case FeSt_NET_SESSION: /**< Network session selection screen, where list of games is displayed, with possibility to join or create own game. */
-        case FeSt_NET_START: /**< Network game start screen (the menu with chat), when created new session or joined existing session. */
-        case FeSt_LEVEL_STATS:
-        case FeSt_HIGH_SCORES:
-        case FeSt_FEDEFINE_KEYS:
-        case FeSt_FEOPTIONS:
-        case FeSt_LEVEL_SELECT:
-        case FeSt_CAMPAIGN_SELECT:
-        case FeSt_MAPPACK_SELECT:
-        case FeSt_MP_MAPPACK_SELECT:
-        case FeSt_LAND_VIEW:
-        case FeSt_NETLAND_VIEW:
-        case FeSt_TORTURE:
-            return true;
-        default:
-            return false;
-    }
-}
+
 
 TbBool all_dungeons_destroyed(const struct PlayerInfo *win_player)
 {
@@ -2104,661 +2080,6 @@ void update_mouse_light(struct PlayerInfo *player)
     set_mouse_light(player, valid, pos);
 }
 
-void check_players_won(void)
-{
-  SYNCDBG(8,"Starting");
-
-    if (!network_is_active())
-        return;
-
-    struct PlayerInfo* curPlayer;
-    for (PlayerNumber playerIdx = 0; playerIdx < PLAYERS_COUNT; ++playerIdx)
-    {
-        curPlayer = get_player(playerIdx);
-        if (!player_exists(curPlayer) || (curPlayer->is_active != 1) || (curPlayer->victory_state != VicS_Undecided))
-            continue;
-
-        // check if any other player is still alive
-        TbBool LivingOpponent = false;
-        for (PlayerNumber secondPlayerIdx = 0; secondPlayerIdx < PLAYERS_COUNT; ++secondPlayerIdx)
-        {
-            if (secondPlayerIdx == playerIdx)
-                continue;
-
-            struct PlayerInfo* otherPlayer = get_player(secondPlayerIdx);
-            if (player_exists(otherPlayer) && otherPlayer->victory_state == VicS_Undecided)
-            {
-                struct Thing* heartng = get_player_soul_container(secondPlayerIdx);
-                if (heartng->active_state != ObSt_BeingDestroyed)
-                {
-                    LivingOpponent = true;
-                    break;
-                }
-            }
-        }
-        if (LivingOpponent == false)
-        {
-            set_player_as_won_level(curPlayer);
-            return;
-        }
-    }
-}
-
-void check_players_lost(void)
-{
-  long i;
-  SYNCDBG(8,"Starting");
-  struct PlayerInfo* player;
-  struct Dungeon* dungeon;
-  for (i=0; i < PLAYERS_COUNT; i++)
-  {
-      player = get_player(i);
-      dungeon = get_players_dungeon(player);
-      if (player_exists(player) && (player->is_active == 1))
-      {
-          struct Thing *heartng;
-          heartng = get_player_soul_container(i);
-          if (heartng->owner != i)
-          {
-              init_player_start(player, true);
-              if (dungeon->dnheart_idx == 0)
-              {
-                  initialise_devastate_dungeon_from_heart(player->id_number);
-              }
-          }
-          if ((!thing_exists(heartng) || ((heartng->active_state == ObSt_BeingDestroyed) && !(dungeon->backup_heart_idx > 0))) && (player->victory_state == VicS_Undecided))
-          {
-            event_kill_all_players_events(i);
-            set_player_as_lost_level(player);
-            //this would easily prevent computer player activities on dead player, but it also makes dead player unable to use
-            //floating spirit, so it can't be done this way: player->is_active = 0;
-            if (is_my_player_number(i)) {
-                LbPaletteSet(engine_palette);
-            }
-          }
-      }
-  }
-}
-
-void blast_slab(MapSlabCoord slb_x, MapSlabCoord slb_y, PlayerNumber plyr_idx)
-{
-    struct SlabMap *slb;
-    slb = get_slabmap_block(slb_x, slb_y);
-    if (slabmap_block_invalid(slb)) {
-        return;
-    }
-    if (slabmap_owner(slb) != plyr_idx) {
-        return;
-    }
-    struct Thing *doortng;
-    doortng = get_door_for_position(slab_subtile_center(slb_x), slab_subtile_center(slb_y));
-    if (!thing_is_invalid(doortng)) {
-        destroy_door(doortng);
-    }
-    struct SlabConfigStats *slabst;
-    slabst = get_slab_stats(slb);
-    if (slabst->category == SlbAtCtg_FortifiedGround)
-    {
-      place_slab_type_on_map(SlbT_PATH, slab_subtile_center(slb_x), slab_subtile_center(slb_y), game.neutral_player_num, 1);
-      decrease_dungeon_area(plyr_idx, 1);
-      do_unprettying(game.neutral_player_num, slb_x, slb_y);
-      do_slab_efficiency_alteration(slb_x, slb_y);
-      struct Coord3d pos;
-      pos.x.val = subtile_coord_center(slab_subtile_center(slb_x));
-      pos.y.val = subtile_coord_center(slab_subtile_center(slb_y));
-      pos.z.val = get_floor_height_at(&pos);
-      create_effect_element(&pos, TngEffElm_RedFlameBig, plyr_idx);
-    }
-}
-
-static void process_dungeon_devastation_effects(void)
-{
-    SYNCDBG(8,"Starting");
-    int plyr_idx;
-    for (plyr_idx=0; plyr_idx < PLAYERS_COUNT; plyr_idx++)
-    {
-        struct Dungeon *dungeon;
-        dungeon = get_players_num_dungeon(plyr_idx);
-        if (dungeon->devastation_turn == 0)
-            continue;
-        if ((get_gameturn() & 1) != 0)
-            continue;
-        dungeon->devastation_turn++;
-        if (dungeon->devastation_turn >= max(game.map_tiles_x,game.map_tiles_y))
-            continue;
-        MapSlabCoord slb_x;
-        MapSlabCoord slb_y;
-        int i;
-        int range;
-        slb_x = subtile_slab(dungeon->devastation_centr_x) - dungeon->devastation_turn;
-        slb_y = subtile_slab(dungeon->devastation_centr_y) - dungeon->devastation_turn;
-        range = 2*dungeon->devastation_turn;
-        for (i = 0; i <= range; i++)
-        {
-            blast_slab(slb_x + i, slb_y,         dungeon->owner);
-            blast_slab(slb_x + i, slb_y + range, dungeon->owner);
-        }
-        for (i = 0; i <= range; i++)
-        {
-            blast_slab(slb_x,         slb_y + i, dungeon->owner);
-            blast_slab(slb_x + range, slb_y + i, dungeon->owner);
-        }
-    }
-}
-
-/**
- * Increments paydays_owed for all players creatures
- * returns amount of creatures needing payday for player
- */
-int set_players_creatures_to_get_paid(PlayerNumber plyr_idx)
-{
-    unsigned long k;
-    long i;
-    int count = 0;
-    const struct StructureList *slist;
-    slist = get_list_for_thing_class(TCls_Creature);
-    i = slist->index;
-    k = 0;
-    while (i != 0)
-    {
-        struct Thing *thing;
-        thing = thing_get(i);
-        if (thing_is_invalid(thing))
-        {
-            ERRORLOG("Jump to invalid thing detected");
-            break;
-        }
-        i = thing->next_of_class;
-        // Per-thing code
-        if (thing->owner == plyr_idx)
-        {
-            struct CreatureModelConfig *crconf;
-            crconf = creature_stats_get_from_thing(thing);
-            if (crconf->pay != 0)
-            {
-                struct CreatureControl *cctrl;
-                cctrl = creature_control_get_from_thing(thing);
-                if (cctrl->paydays_advanced > 0)
-                {
-                    cctrl->paydays_advanced--;
-                } else
-                {
-                    if (!creature_is_kept_in_custody_by_enemy(thing))
-                    {
-                        cctrl->paydays_owed++;
-                        count++;
-                    }
-                    else
-                    {
-                        cctrl->paydays_advanced--;
-                    }
-                }
-            }
-        }
-        // Per-thing code ends
-        k++;
-        if (k > THINGS_COUNT)
-        {
-            ERRORLOG("Infinite loop detected when sweeping things list");
-            break;
-        }
-    }
-    return count;
-}
-
-void process_payday(void)
-{
-    PlayerNumber plyr_idx;
-    for (plyr_idx=0; plyr_idx < PLAYERS_COUNT; plyr_idx++)
-    {
-        game.pay_day_progress[plyr_idx] = game.pay_day_progress[plyr_idx] + (game.conf.rules[plyr_idx].gameplay.pay_day_speed / 100);
-        if (player_is_roaming(plyr_idx) || (plyr_idx == game.neutral_player_num)) {
-            continue;
-        }
-        struct PlayerInfo *player;
-        player = get_player(plyr_idx);
-        if (player_exists(player) && (player->is_active == 1))
-        {
-            compute_and_update_player_payday_total(plyr_idx);
-            compute_and_update_player_backpay_total(plyr_idx);
-        }
-    }
-    int player_paid_creatures_count;
-    for (plyr_idx = 0; plyr_idx < PLAYERS_COUNT; plyr_idx++)
-    {
-        if (game.conf.rules[plyr_idx].gameplay.pay_day_gap <= game.pay_day_progress[plyr_idx])
-        {
-            if (is_my_player_number(plyr_idx))
-                output_message(SMsg_Payday, 0);
-            game.pay_day_progress[plyr_idx] = 0;
-            player_paid_creatures_count = set_players_creatures_to_get_paid(plyr_idx);
-            if (player_paid_creatures_count > 0)
-            {
-                struct Dungeon *dungeon = get_players_num_dungeon(plyr_idx);
-                event_create_event_or_update_nearby_existing_event(0, 0, EvKind_CreaturePayday, plyr_idx, dungeon->creatures_total_pay);
-            }
-        }
-    }
-}
-
-void process_dungeons(void)
-{
-  SYNCDBG(7,"Starting");
-  check_players_won();
-  check_players_lost();
-  process_dungeon_power_magic();
-  process_dungeon_devastation_effects();
-  process_entrance_generation();
-  process_payday();
-  process_things_in_dungeon_hand();
-  SYNCDBG(9,"Finished");
-}
-
-void update_near_creatures_for_footsteps(int32_t *near_creatures, const struct Coord3d *srcpos)
-{
-    long near_distance[3];
-    // Don't allow creatures which are far by over 20 subtiles
-    near_distance[0] = subtile_coord(20,0);
-    near_distance[1] = subtile_coord(20,0);
-    near_distance[2] = subtile_coord(20,0);
-    near_creatures[0] = 0;
-    near_creatures[1] = 0;
-    near_creatures[2] = 0;
-    // Find the closest thing for footsteps
-    struct Thing *thing;
-    unsigned long k;
-    long i;
-    const struct StructureList *slist;
-    slist = get_list_for_thing_class(TCls_Creature);
-    i = slist->index;
-    k = 0;
-    while (i != 0)
-    {
-        thing = thing_get(i);
-        if (thing_is_invalid(thing))
-        {
-            ERRORLOG("Jump to invalid thing detected");
-            break;
-        }
-        i = thing->next_of_class;
-        // Per-thing code
-        thing->state_flags &= ~TF1_DoFootsteps;
-        if ( (!thing_is_picked_up(thing)) && (!thing_is_dragged_or_pulled(thing)) )
-        {
-            struct CreatureSound *crsound;
-            crsound = get_creature_sound(thing, CrSnd_Foot);
-            if (crsound->index != 0)
-            {
-                struct CreatureControl *cctrl;
-                cctrl = creature_control_get_from_thing(thing);
-                long ndist;
-                ndist = get_chessboard_distance(srcpos, &thing->mappos);
-                if (ndist < near_distance[0])
-                {
-                    if (((cctrl->distance_to_destination != 0) && ((int)thing->floor_height >= (int)thing->mappos.z.val))
-                      || ((thing->movement_flags & TMvF_Flying) != 0))
-                    {
-                        // Insert the new item to our list
-                        int n;
-                        for (n = 2; n>0; n--)
-                        {
-                            near_creatures[n] = near_creatures[n-1];
-                            near_distance[n] = near_distance[n-1];
-                        }
-                        near_distance[0] = ndist;
-                        near_creatures[0] = thing->index;
-                    }
-                }
-            }
-        }
-        // Per-thing code ends
-        k++;
-        if (k > THINGS_COUNT)
-        {
-            ERRORLOG("Infinite loop detected when sweeping things list");
-            break;
-        }
-    }
-}
-
-long stop_playing_flight_sample_in_all_flying_creatures(void)
-{
-    struct Thing *thing;
-    unsigned long k;
-    long i;
-    long naffected;
-    naffected = 0;
-    const struct StructureList *slist;
-    slist = get_list_for_thing_class(TCls_Creature);
-    i = slist->index;
-    k = 0;
-    while (i != 0)
-    {
-        thing = thing_get(i);
-        if (thing_is_invalid(thing))
-        {
-          ERRORLOG("Jump to invalid thing detected");
-          break;
-        }
-        i = thing->next_of_class;
-        // Per-thing code
-        if ((get_creature_model_flags(thing) & CMF_IsDiptera) && ((thing->state_flags & TF1_DoFootsteps) == 0))
-        {
-            if ( S3DEmitterIsPlayingSample(thing->snd_emitter_id, 25) ) {
-                S3DDeleteSampleFromEmitter(thing->snd_emitter_id, 25);
-            }
-        }
-        // Per-thing code ends
-        k++;
-        if (k > THINGS_COUNT)
-        {
-          ERRORLOG("Infinite loop detected when sweeping things list");
-          break;
-        }
-    }
-    return naffected;
-}
-
-void update_footsteps_nearest_camera(struct Camera *cam)
-{
-    static long timeslice = 0;
-    static int32_t near_creatures[3];
-    struct Coord3d srcpos;
-    SYNCDBG(6,"Starting");
-    if (cam == NULL)
-        return;
-    srcpos.x.val = cam->mappos.x.val;
-    srcpos.y.val = cam->mappos.y.val;
-    srcpos.z.val = cam->mappos.z.val;
-    if (timeslice == 0) {
-        update_near_creatures_for_footsteps(near_creatures, &srcpos);
-    }
-    long i;
-    for (i=0; i < 3; i++)
-    {
-        struct Thing *thing;
-        if (near_creatures[i] == 0)
-            break;
-        thing = thing_get(near_creatures[i]);
-        if (thing_is_creature(thing)) {
-            thing->state_flags |= TF1_DoFootsteps;
-            play_thing_walking(thing);
-        }
-    }
-    if (timeslice == 0)
-    {
-        stop_playing_flight_sample_in_all_flying_creatures();
-    }
-    timeslice = (timeslice + 1) % 4;
-}
-
-int clear_active_dungeons_stats(void)
-{
-  struct Dungeon *dungeon;
-  int i;
-  for (i=0; i < PLAYERS_COUNT; i++)
-  {
-      dungeon = get_dungeon(i);
-      if (dungeon_invalid(dungeon))
-          break;
-      memset((char *)dungeon->crmodel_state_type_count, 0, game.conf.crtr_conf.model_count * STATE_TYPES_COUNT * sizeof(uint16_t));
-      memset((char *)dungeon->guijob_all_creatrs_count, 0, game.conf.crtr_conf.model_count *3*sizeof(uint16_t));
-      memset((char *)dungeon->guijob_angry_creatrs_count, 0, game.conf.crtr_conf.model_count *3*sizeof(uint16_t));
-  }
-  return i;
-}
-
-TngUpdateRet damage_creatures_with_physical_force(struct Thing *thing, ModTngFilterParam param)
-{
-    SYNCDBG(18,"Starting for %s index %d",thing_model_name(thing),(int)thing->index);
-    if (thing_is_picked_up(thing) || thing_is_dragged_or_pulled(thing))
-    {
-        return TUFRet_Unchanged;
-    }
-    if (thing_is_creature(thing))
-    {
-        apply_damage_to_thing_and_display_health(thing, param->secondary_number, param->primary_number);
-        if ((thing->health >= 0) && !creature_is_leaving_and_cannot_be_stopped(thing))
-        {
-            if (((thing->alloc_flags & TAlF_IsControlled) == 0) && !creature_is_kept_in_custody(thing))
-            {
-                if (get_creature_state_besides_interruptions(thing) != CrSt_CreatureEscapingDeath)
-                {
-                    if (cleanup_current_thing_state(thing) && setup_move_out_of_cave_in(thing))
-                        thing->continue_state = CrSt_CreatureEscapingDeath;
-                }
-            }
-            return TUFRet_Modified;
-        } else
-        {
-            kill_creature(thing, INVALID_THING, param->primary_number, CrDed_NoEffects|CrDed_DiedInBattle);
-            return TUFRet_Deleted;
-        }
-    }
-    else if (thing_is_destructible_trap(thing) > 0)
-    {
-        apply_damage_to_thing(thing, param->secondary_number, param->primary_number);
-        return TUFRet_Modified;
-    }
-    return TUFRet_Unchanged;
-}
-
-TbBool valid_cave_in_position(PlayerNumber plyr_idx, MapSubtlCoord stl_x, MapSubtlCoord stl_y)
-{
-    struct Map *mapblk;
-    mapblk = get_map_block_at(stl_x,stl_y);
-    if ((mapblk->flags & SlbAtFlg_Blocking) != 0)
-        return false;
-    struct SlabMap *slb;
-    slb = get_slabmap_for_subtile(stl_x,stl_y);
-    return (plyr_idx == game.neutral_player_num) || (slabmap_owner(slb) == game.neutral_player_num) || (slabmap_owner(slb) == plyr_idx);
-}
-
-long update_cave_in(struct Thing *thing)
-{
-    thing->health--;
-    thing->rendering_flags |= TRF_Invisible;
-    if (thing->health < 1)
-    {
-        delete_thing_structure(thing, 0);
-        return 1;
-    }
-
-    const struct PowerConfigStats *powerst;
-    powerst = get_power_model_stats(PwrK_CAVEIN);
-    struct Thing *efftng;
-    struct Coord3d pos;
-    PlayerNumber owner;
-    owner = thing->owner;
-    if ((get_gameturn() % 3) == 0)
-    {
-        int n;
-        n = GAME_RANDOM(AROUND_TILES_COUNT);
-        pos.x.val = thing->mappos.x.val + GAME_RANDOM(704) * around[n].delta_x;
-        pos.y.val = thing->mappos.y.val + GAME_RANDOM(704) * around[n].delta_y;
-        if (subtile_has_slab(coord_subtile(pos.x.val),coord_subtile(pos.y.val)))
-        {
-            pos.z.val = get_ceiling_height(&pos) - 128;
-            efftng = create_effect_element(&pos, TngEff_Flash, owner);
-            if (!thing_is_invalid(efftng)) {
-                efftng->health = powerst->duration;
-            }
-        }
-    }
-
-    GameTurnDelta turns_between;
-    GameTurnDelta turns_alive;
-    turns_between = powerst->duration / 5;
-    turns_alive = get_gameturn() - thing->creation_turn;
-    if ((turns_alive != 0) && ((turns_between < 1) || (3 * turns_between / 4 == turns_alive % turns_between)))
-    {
-        pos.x.val = thing->mappos.x.val + THING_RANDOM(thing, 128);
-        pos.y.val = thing->mappos.y.val + THING_RANDOM(thing, 128);
-        pos.z.val = get_floor_height_at(&pos) + 384;
-        create_effect(&pos, TngEff_HarmlessGas4, owner);
-    }
-
-    if ((turns_alive % game.conf.rules[owner].magic.turns_per_collapse_dngn_dmg) == 0)
-    {
-        pos.x.val = thing->mappos.x.val;
-        pos.y.val = thing->mappos.y.val;
-        pos.z.val = subtile_coord(1,0);
-        Thing_Modifier_Func do_cb;
-        struct CompoundTngFilterParam param;
-        param.plyr_idx = -1;
-        param.class_id = 0;
-        param.model_id = 0;
-        param.primary_number = thing->owner;
-        param.secondary_number = game.conf.rules[thing->owner].magic.collapse_dungeon_damage;
-        param.tertiary_pointer = 0;
-        do_cb = damage_creatures_with_physical_force;
-        do_to_things_with_param_around_map_block(&pos, do_cb, &param);
-    }
-
-    if ((8 * powerst->duration / 10 >= thing->health) && (2 * powerst->duration / 10 <= thing->health))
-    {
-        if ((powerst->duration < 10) || ((thing->health % (powerst->duration / 10)) == 0))
-        {
-            int round_idx;
-            round_idx = THING_RANDOM(thing, AROUND_TILES_COUNT);
-            set_coords_to_slab_center(&pos, subtile_slab(thing->mappos.x.val + 3 * around[round_idx].delta_x), subtile_slab(thing->mappos.y.val + 3 * around[round_idx].delta_y));
-            if (subtile_has_slab(coord_subtile(pos.x.val), coord_subtile(pos.y.val)) && valid_cave_in_position(thing->owner, coord_subtile(pos.x.val), coord_subtile(pos.y.val)))
-            {
-                struct Thing *ncavitng;
-                ncavitng = get_cavein_at_subtile_owned_by(coord_subtile(pos.x.val), coord_subtile(pos.y.val), -1);
-                if (thing_is_invalid(ncavitng))
-                {
-                    long dist;
-                    struct Coord3d pos2;
-                    pos2.x.val = subtile_coord(thing->cave_in.x,0);
-                    pos2.y.val = subtile_coord(thing->cave_in.y,0);
-                    pos2.z.val = subtile_coord(1,0);
-                    dist = get_chessboard_distance(&pos, &pos2);
-                    if (powerst->strength[thing->cave_in.model] >= coord_subtile(dist))
-                    {
-                        ncavitng = create_thing(&pos, TCls_CaveIn, thing->cave_in.model, owner, -1);
-                        if (!thing_is_invalid(ncavitng))
-                        {
-                            thing->health += 5;
-                            if (thing->health > 0)
-                            {
-                                ncavitng->cave_in.x = thing->cave_in.x;
-                                ncavitng->cave_in.y = thing->cave_in.y;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-    return 1;
-}
-
-/**
- * rules can change by dkscript/lua.
- * Checks if a gamerule for lighting has changed and updates the lights if they are.
- * This function also refreshes the light status of the map.
-*/
-void update_global_lighting()
-{
-    if (!game.lish.light_auto_sync)
-        return;
-
-    // Check if any values have changed
-    if (
-        game.conf.rules[0].gameplay.global_ambient_light != game.lish.global_ambient_light ||
-        game.conf.rules[0].gameplay.light_enabled != game.lish.light_enabled
-    ){
-
-        // GlobalAmbientLight
-        if (game.conf.rules[0].gameplay.global_ambient_light != game.lish.global_ambient_light)
-        {
-            game.lish.global_ambient_light = game.conf.rules[0].gameplay.global_ambient_light;
-        }
-
-        // LightEnabled
-        if (game.conf.rules[0].gameplay.light_enabled != game.lish.light_enabled)
-        {
-            game.lish.light_enabled = game.conf.rules[0].gameplay.light_enabled;
-        }
-
-        // Refresh the lights
-        light_stat_refresh();
-    }
-}
-
-void update(void)
-{
-    struct PlayerInfo *player;
-    SYNCDBG(4,"Starting for turn %ld",(long)get_gameturn());
-
-    update_local_cameras_pre();
-    process_packets();
-    api_update_server();
-
-    if (quit_game || exit_keeper) {
-        return;
-    }
-    if (game.game_kind == GKind_NonInteractiveState)
-    {
-        game.map_changed_for_navigation = 0;
-        return;
-    }
-    player = get_my_player();
-
-    if (!flag_is_set(game.operation_flags,GOF_Paused))
-    {
-        if (flag_is_set(player->additional_flags,PlaAF_LightningPaletteIsActive))
-        {
-            PaletteSetPlayerPalette(player, engine_palette);
-            clear_flag(player->additional_flags, PlaAF_LightningPaletteIsActive);
-        }
-        clear_active_dungeons_stats();
-        update_creature_pool_state();
-        if ((get_gameturn() & 0x01) != 0)
-            update_animating_texture_maps();
-        update_things();
-        process_rooms();
-        process_dungeons();
-        update_research();
-        update_manufacturing();
-        event_process_events();
-        update_all_events();
-        process_level_script();
-        process_fx_lines();
-        lua_on_game_tick();
-        if ((game.view_mode_flags & GNFldD_ComputerPlayerProcessing) != 0)
-            process_computer_players2();
-        process_players();
-        process_action_points();
-        player = get_my_player();
-        if (player->view_mode == PVM_CreatureView)
-        {
-            struct Thing *thing = thing_get(player->controlled_thing_idx);
-            update_first_person_object_ambience(thing);
-        }
-        update_footsteps_nearest_camera(get_player_active_camera(player));
-        PaletteFadePlayer(player);
-        process_armageddon();
-        update_global_lighting();
-#if (BFDEBUG_LEVEL > 9)
-        lights_stats_debug_dump();
-        things_stats_debug_dump();
-        creature_stats_debug_dump();
-#endif
-        game.play_gameturn++;
-        if (game.turns_packetoff == game.play_gameturn)
-            exit_keeper = 1;
-    }
-
-    message_update();
-    update_all_players_cameras();
-    update_local_cameras_post();
-    update_player_sounds();
-    SYNCDBG(6,"Finished");
-}
-
-
 long near_map_block_thing_filter_queryable_object(const struct Thing *thing, MaxTngFilterParam param, long maximizer)
 {
 /* Currently this only makes Dungeon Heart blinking; maybe I'll find a purpose for it later
@@ -3192,134 +2513,10 @@ void engine(struct PlayerInfo *player, struct Camera *cam)
     LbScreenLoadGraphicsWindow(&grwnd);
 }
 
-void find_frame_rate(void)
-{
-    static TbClockMSec prev_time2=0;
-    static TbClockMSec cntr_time2=0;
-    unsigned long curr_time;
-    curr_time = LbTimerClock();
-    cntr_time2++;
-    if (curr_time-prev_time2 >= 1000)
-    {
-        double time_fdelta = 1000.0*((double)(cntr_time2))/(curr_time-prev_time2);
-        prev_time2 = curr_time;
-        game.time_delta = (unsigned long)(time_fdelta*256.0);
-        cntr_time2 = 0;
-    }
-}
 
-void packet_load_find_frame_rate(unsigned long incr)
-{
-    static TbClockMSec start_time=0;
-    static TbClockMSec extra_frames=0;
-    TbClockMSec curr_time;
-    curr_time = LbTimerClock();
-    if ((curr_time-start_time) < 5000)
-    {
-        extra_frames += incr;
-    } else
-    {
-        double time_fdelta = 1000.0*((double)(extra_frames+incr))/(curr_time-start_time);
-        start_time = curr_time;
-        game.time_delta = (unsigned long)(time_fdelta*256.0);
-        extra_frames = 0;
-    }
-}
 
-/**
- * Checks if the game screen needs redrawing.
- */
-short display_should_be_updated_this_turn(void)
-{
-    if ((game.operation_flags & GOF_Paused) != 0)
-      return true;
-    if ( (game.turns_fastforward == 0) && (!game.packet_loading_in_progress) )
-    {
-      find_frame_rate();
-      if ( (game.frame_skip == 0) || ((get_gameturn() % game.frame_skip) == 0) )
-        return true;
-    } else
-    if ( ((get_gameturn() & 0x3F)==0) ||
-         ((game.packet_loading_in_progress) && ((get_gameturn() & 7)==0)) )
-    {
-      packet_load_find_frame_rate(64);
-      return true;
-    }
-    return false;
-}
 
-/**
- * Makes last updates to the video buffer, and swaps buffers to show
- * the new image.
- */
-TbBool keeper_screen_swap(void)
-{
-/*  // For resolution 640x480, move the graphics data 40 lines lower
-  if ( lbDisplay.ScreenMode == Lb_SCREEN_MODE_640_480_8 )
-    if (LbScreenLock() == Lb_SUCCESS)
-    {
-      int i;
-      int scrmove_x=0;
-      int scrmove_y=40;
-      int scanline_len=640;
-      for (i=400;i>=0;i--)
-        memcpy(lbDisplay.WScreen+scanline_len*(i+scrmove_y)+scrmove_x, lbDisplay.WScreen+scanline_len*i, scanline_len-scrmove_x);
-      memset(lbDisplay.WScreen, 0, scanline_len*scrmove_y);
-      LbScreenUnlock();
-    }*/
-  LbScreenSwap();
-  return true;
-}
 
-/**
- * Waits until the next game turn. Delay is usually controlled by
- * num_fps variable.
- */
-extern "C" {
-int32_t multiplayer_speed_adjustment_ns;
-}
-
-TbBool keeper_wait_for_next_turn(void)
-{
-    const long double tick_ns_one_sec = 1000000000.0;
-    long double tick_ns_one_frame = -1;
-    if ((game.view_mode_flags & GNFldD_WaitSleepMode) != 0)
-    {
-        // No idea when such situation occurs
-        tick_ns_one_frame = tick_ns_one_sec;
-    }
-    if (game.frame_skip >= 0)
-    {
-        // Standard delaying system
-        int32_t num_fps = turns_per_second;
-        if (game.frame_skip > 0)
-            num_fps *= game.frame_skip;
-
-        tick_ns_one_frame = tick_ns_one_sec/num_fps;
-    }
-
-    if (tick_ns_one_frame >= 0) {
-        static long double tick_ns_last_turn = 0;
-
-        long double tick_ns_cur = get_time_tick_ns();
-        long double tick_ns_used = tick_ns_cur - tick_ns_last_turn;
-        long double tick_ns_delay = tick_ns_one_frame - tick_ns_used;
-        if (multiplayer_speed_adjustment_ns != 0) {
-            tick_ns_delay += multiplayer_speed_adjustment_ns;
-        }
-
-        long double tick_ns_end = tick_ns_cur;
-        // tick_ns_used: every level, initialized_time_point will be reset, so tick_ns_used may be less than 0 when enter level for the non-first time, Skip it directly to solve the problem.
-        if (tick_ns_delay > 0 && tick_ns_used >= 0) {
-            tick_ns_end = tick_ns_cur + tick_ns_delay;
-            LbSleepUntilExt(tick_ns_end);
-        }
-        tick_ns_last_turn = tick_ns_end;
-        return true;
-    }
-
-    return false;
-}
 
 void redetect_screen_refresh_rate_for_draw()
 {
@@ -3344,13 +2541,13 @@ void redetect_screen_refresh_rate_for_draw()
     }
 }
 
-static bool use_delta_time()
+bool use_delta_time()
 {
     // Always enable interpolation in multiplayer games.
     return is_feature_on(Ft_DeltaTime) || network_is_active();
 }
 
-static void update_frontend_delta_time()
+void update_frontend_delta_time()
 {
     static int64_t prev = 0;
     const int64_t now = get_time_tick_ns();
@@ -3360,7 +2557,7 @@ static void update_frontend_delta_time()
     game.delta_time = min(max(dt, 0.L), 1.L);
 }
 
-static void update_gameplay_delta_time()
+void update_gameplay_delta_time()
 {
     if (use_delta_time()) {
         static int64_t prev = 0;
@@ -3390,222 +2587,7 @@ static void update_gameplay_delta_time()
     }
 }
 
-static bool keeper_wait_for_screen_focus()
-{
-    do {
-        if ( !poll_inputs() )
-        {
-          force_application_close();
-          break;
-        }
-        if (LbIsActive())
-          return true;
-        if (network_is_active())
-          return true;
-        if (!freeze_game_on_focus_lost())
-          return true;
-        LbSleepFor(50);
-        update_gameplay_delta_time();
-        game.process_turn_time = 1.0;
-        time_since_last_draw = 1.0;
-    } while ((!exit_keeper) && (!quit_game));
-    return false;
-}
-
-static void gameplay_loop_draw()
-{
-    if (use_delta_time())
-        do_draw = true;
-
-    update_gameplay_delta_time();
-
-    if (game.process_turn_time > 1.0 && time_since_last_draw < 1.0)
-        do_draw = false;
-
-    // Frame rate limiter
-    if (fps_limit_current > 0)
-    {
-        frametime_start_measurement(Frametime_Sleep);
-        if (process_frame_time < 1.0)
-        {
-            if (game.process_turn_time < 1.0)
-                SDL_Delay(1);
-            do_draw = false;
-        }
-        else
-        {
-            process_frame_time = min(1.L, process_frame_time - 1.L);
-        }
-        frametime_end_measurement(Frametime_Sleep);
-    }
-
-    // Floats are used a lot in the drawing related functions. But keep in mind integers are typically preferred for logic related functions.
-    frametime_start_measurement(Frametime_Draw);
-
-    // Update lights
-    update_light_render_area();
-
-    if (quit_game || exit_keeper) {
-        do_draw = false;
-    }
-    if ( do_draw ) {
-        if (frametime_enabled())
-            framerate_measurement_capture(Framerate_Draw);
-        game.delta_time = min(time_since_last_draw, 1.L);
-        time_since_last_draw = 0;
-        interpolate_time = min(max(game.process_turn_time, 0.L), 1.L);
-        keeper_screen_redraw();
-    }
-    keeper_wait_for_screen_focus();
-    // Direct information/error messages
-    if (LbScreenLock() == Lb_SUCCESS) {
-        if ( do_draw ) {
-            perform_any_screen_capturing();
-        }
-        draw_onscreen_direct_messages();
-        LbScreenUnlock();
-    }
-    // Move the graphics window to center of screen buffer and swap screen
-    if ( do_draw ) {
-        keeper_screen_swap();
-    }
-    frametime_end_measurement(Frametime_Draw);
-
-    if ( do_draw ) {
-        update_gameplay_delta_time();
-        const long double delta = time_since_last_draw - average_frame_draw_time;
-        average_frame_draw_time += delta * max(average_frame_draw_time, .05L) / 20;
-    }
-}
-
-static void gameplay_loop_logic()
-{
-    if(flag_is_set(start_params.debug_flags, DFlg_PauseAtGameTurn))
-    {
-        static GameTurn previous_gameturn = 0;
-        if(get_gameturn() >= start_params.pause_at_gameturn && get_gameturn() != previous_gameturn)
-        {
-            if(!game.paused_at_gameturn)
-            {
-                game.paused_at_gameturn = true;
-
-                game.frame_skip = 0;
-                if(game.packet_load_enable)
-                {
-                    disable_packet_mode();
-                }
-                set_packet_pause_toggle();
-            }
-        }
-        previous_gameturn = get_gameturn();
-    }
-
-    if (use_delta_time())
-    {
-        update_gameplay_delta_time();
-        if (game.input_lag_turns == 0 && network_is_active())
-        {
-            // Aim to exchange network packets before the turn ends.  If drawing
-            // another frame could miss this deadline, skip it.
-            // In a 3-4 player game, clients must be 2 frames early.
-            const int frames = 1 + (netstate.my_id != SERVER_ID && game.active_players_count > 2);
-            const long double offset = frames * average_frame_draw_time * multiplayer_clock_adjust * max(game.frame_skip, 1);
-            if (game.process_turn_time + offset < 1.0)
-                return;
-        }
-        else
-        {
-            if (game.process_turn_time < 1.0)
-                return;
-        }
-    }
-
-    frametime_start_measurement(Frametime_Logic);
-    if (frametime_enabled())
-        framerate_measurement_capture(Framerate_Logic);
-
-#ifdef FUNCTESTING
-    if(flag_is_set(start_params.functest_flags, FTF_Enabled))
-    {
-        FTestFrameworkState ftstate = ftest_update(NULL);
-        if(ftstate == FTSt_InvalidState || ftstate == FTSt_TestsCompletedSuccessfully)
-        {
-            quit_game = true;
-            exit_keeper = true;
-            return;
-        }
-    }
-#endif // FUNCTESTING
-    do_draw = display_should_be_updated_this_turn() || (!LbIsActive());
-    poll_inputs();
-    input_eastegg();
-    input();
-    exchange_packets();
-
-    update_gameplay_delta_time();
-    if (game.process_turn_time > turns_per_second + 1)
-        game.process_turn_time = turns_per_second + 1;
-
-    // Adjust client time scaling
-    if (netstate.my_id != SERVER_ID && network_is_active())
-    {
-        if (game.input_lag_turns == 0)
-        {
-            // Adjust the clock rate so that the host packet is received at
-            // process_turn_time == 1.0 (on average).  If it is received later,
-            // reduce the scaling factor (< 1.0) so that the next turn takes a
-            // little longer in real time.  Vice-versa if it is early.
-
-            multiplayer_clock_adjust = 1 + (1 - host_packet_received) / 20;
-        }
-        else
-        {
-            const long double tick_ns_one_turn = 1e9L / turns_per_second;
-            const long double tick_ns_adjusted_turn = tick_ns_one_turn + multiplayer_speed_adjustment_ns;
-            assert (tick_ns_adjusted_turn > 0);
-            multiplayer_clock_adjust = tick_ns_one_turn / tick_ns_adjusted_turn;
-        }
-    }
-    else multiplayer_clock_adjust = 1.0;
-    host_packet_received = 1.0;
-
-    while (game.process_turn_time < 1.0)
-    {
-        gameplay_loop_draw();
-        update_gameplay_delta_time();
-    }
-    game.process_turn_time -= 1.0;
-
-    update();
-
-    frametime_end_measurement(Frametime_Logic);
-
-    if(game.frame_step)
-    {
-        game.frame_step = false;
-        set_packet_pause_toggle();
-    }
-}
-
-static void gameplay_loop_network()
-{
-    if (! network_is_active())
-        return;
-
-    network_update(game.packets, sizeof(struct Packet));
-}
-
-static void gameplay_loop_timestep()
-{
-    if (! use_delta_time()) {
-        frametime_start_measurement(Frametime_Sleep);
-        // Make delay if the machine is too fast
-        if ( (!game.packet_load_enable) || (game.turns_fastforward == 0) ) {
-            keeper_wait_for_next_turn();
-        }
-        frametime_end_measurement(Frametime_Sleep);
-    }
-}
+void gameplay_loop_draw();
 
 extern "C" void network_yield_draw_gameplay()
 {
@@ -3644,41 +2626,7 @@ extern "C" void network_yield_draw_frontend()
     LbScreenSwap();
 }
 
-void keeper_gameplay_loop(void)
-{
-    struct PlayerInfo *player;
-    SYNCDBG(5,"Starting");
-    player = get_my_player();
-    PaletteSetPlayerPalette(player, engine_palette);
-    if ((game.operation_flags & GOF_SingleLevel) != 0) {
-        initialise_eye_lenses();
-    }
-    SYNCDBG(0,"Entering the gameplay loop for level %d",(int)get_loaded_level_number());
-    LbErrorParachuteUpdate(); // For some reasone parachute keeps changing; Remove when won't be needed anymore
 
-    initial_time_point();
-    LbSleepExtInit();
-
-    //the main gameplay loop starts
-    while ((!quit_game) && (!exit_keeper))
-    {
-        frametime_start_measurement(Frametime_FullFrame);
-        if (frametime_enabled())
-            framerate_measurement_capture(Framerate_FullFrame);
-        gameplay_loop_logic();
-        gameplay_loop_draw();
-        gameplay_loop_network();
-        gameplay_loop_timestep();
-
-        frametime_end_measurement(Frametime_FullFrame);
-    } // end while
-    SYNCDBG(0,"Gameplay loop finished after %lu turns",(unsigned long)get_gameturn());
-
-    // Reset the game kind because we are not in a game anymore at this point
-    game.game_kind = GKind_Unset;
-
-    api_event("GAME_ENDED");
-}
 
 TbBool can_thing_be_queried(struct Thing *thing, PlayerNumber plyr_idx)
 {
@@ -3763,395 +2711,6 @@ void initialise_map_health(void)
     }
 }
 
-static TbBool wait_at_frontend(void)
-{
-    struct PlayerInfo *player;
-    // This is an improvised coroutine-like stuff
-    CoroutineLoop loop;
-    memset(&loop, 0, sizeof(loop));
-
-    SYNCDBG(0,"Falling into frontend menu.");
-    // Moon phase calculation
-    calculate_moon_phase(true,false);
-    update_extra_levels_visibility();
-    // Returning from Demo Mode
-    if (game.mode_flags & MFlg_IsDemoMode)
-    {
-      close_packet_file();
-      game.packet_load_enable = 0;
-    }
-    game.save_game_slot = -1;
-    // Make sure campaigns are loaded
-    if (!load_campaigns_list(&campaigns_list ,FGrp_Campgn ,"campaigns","campgn_order.txt"))
-    {
-      ERRORLOG("No valid campaign files found");
-      exit_keeper = 1;
-      return true;
-    }
-    // Make sure mappacks are loaded
-    if (!load_campaigns_list(&mappacks_list,FGrp_VarLevels,"mappacks","mappck_order.txt"))
-    {
-      WARNMSG("No valid mappack files found");
-    }
-    if (!load_campaigns_list(&mp_mappacks_list,FGrp_MpLevels,"multiplayer mappacks","mp_mappck_order.txt"))
-    {
-      WARNMSG("No valid multiplayer mappack files found");
-    }
-    //Set level number and campaign (for single level mode: GOF_SingleLevel)
-    if ((start_params.operation_flags & GOF_SingleLevel) != 0)
-    {
-        TbBool result = false;
-        if (start_params.selected_campaign[0] != '\0')
-        {
-            result = change_campaign(CampgnT_Default, start_params.selected_campaign);
-        }
-        if (!result) {
-            if (!change_campaign(CampgnT_Default,"")) {
-                WARNMSG("Unable to load default campaign for the specified level CMD Line parameter");
-            }
-            else if (start_params.selected_campaign[0] != '\0') { // only show this log message if the user actually specified a campaign
-                WARNMSG("Unable to load campaign associated with the specified level CMD Line parameter, default loaded.");
-            }
-            else {
-                JUSTLOG("No campaign specified. Default campaign loaded for selected level (%u).", start_params.selected_level_number);
-            }
-        }
-        set_selected_level_number(start_params.selected_level_number);
-        //game.selected_level_number = start_params.selected_level_number;
-    }
-    else
-    {
-        set_selected_level_number(first_singleplayer_level());
-    }
-    // Init load/save catalogue
-    initialise_load_game_slots();
-
-    #ifdef FUNCTESTING
-    if(flag_is_set(start_params.functest_flags, FTF_Enabled)) //override for functional tests
-    {
-        FTestFrameworkState ft_prev_state = FTSt_InvalidState;
-        FTestFrameworkState ft_current_state = ftest_update(&ft_prev_state);
-
-        TbBool user_aborted_tests = ft_prev_state == FTSt_TestIsProcessingActions && ft_current_state == FTSt_TestIsProcessingActions;
-        if(user_aborted_tests)
-        {
-            FTEST_FAIL_TEST("User aborted tests");
-        }
-
-        if(ft_current_state == FTSt_InvalidState || ft_current_state == FTSt_TestsCompletedSuccessfully || user_aborted_tests)
-        {
-            quit_game = true;
-            exit_keeper = true;
-            return true;
-        }
-        faststartup_network_game(&loop);
-        coroutine_process(&loop);
-        return true;
-    }
-    #endif
-
-    // Prepare to enter PacketLoad game
-    if ((game.packet_load_enable) && (!game.packet_load_initialized))
-    {
-      faststartup_saved_packet_game();
-      return true;
-    }
-    // Load single-player level directly from command line arguments (-server and -connect bypass this, autoloading a multiplayer map is handled elsewhere)
-    if ((game.operation_flags & GOF_SingleLevel) != 0 && !(game_flags2 & (GF2_Connect | GF2_Server)))
-    {
-      faststartup_network_game(&loop);
-      coroutine_process(&loop);
-      return true;
-    }
-
-    if ( !setup_screen_mode_minimal(get_frontend_vidmode()) )
-    {
-      FatalError = 1;
-      exit_keeper = 1;
-      return true;
-    }
-    LbScreenClear(0);
-    LbScreenSwap();
-    if (frontend_load_data() != Lb_SUCCESS)
-    {
-      ERRORLOG("Unable to load frontend data");
-      exit_keeper = 1;
-      return true;
-    }
-    memset(scratch, 0, PALETTE_SIZE);
-    LbPaletteSet(scratch);
-    frontend_set_state(get_startup_menu_state());
-
-    // Once the Mouse Sprite initialization is complete, the sprite's position needs to be reset because it defaults to (0, 0).
-    // Note that we cannot use LbMoveGameCursorToHostCursor for this, because the buffer position may remain unchanged.
-    LbMouseSetPositionInitial(lbDisplay.MMouseX, lbDisplay.MMouseY);
-
-    try_restore_frontend_error_box();
-
-    poll_inputs();
-    clear_mouse_pressed_lrbutton();
-
-    short finish_menu = 0;
-    clear_flag(game.mode_flags, MFlg_DemoMode);
-    // TODO move to separate function
-    // Begin the frontend loop
-    long fe_last_loop_time = LbTimerClock();
-    do
-    {
-      if (!poll_inputs())
-      {
-        force_application_close();
-        SYNCDBG(0,"Windows Control exit condition invoked");
-        break;
-      }
-      update_mouse();
-      update_key_modifiers();
-      old_mouse_over_button = frontend_mouse_over_button;
-      frontend_mouse_over_button = 0;
-
-      frontend_input();
-      if ( exit_keeper )
-      {
-        SYNCDBG(0,"Frontend Input exit condition invoked");
-        break; // end while
-      }
-
-      frontend_update(&finish_menu);
-      if ( exit_keeper )
-      {
-        SYNCDBG(0,"Frontend Update exit condition invoked");
-        break; // end while
-      }
-
-      if ((!finish_menu) && (LbIsActive()))
-      {
-        frontend_draw();
-        LbScreenSwap();
-      }
-
-      if (!SoundDisabled)
-      {
-        process_3d_sounds();
-        MonitorStreamedSoundTrack();
-      }
-
-      if (fade_palette_in)
-      {
-        fade_in();
-        fade_palette_in = 0;
-      } else {
-        if (is_feature_on(Ft_DeltaTime) == true && should_use_delta_time_on_menu()) {
-          update_frontend_delta_time();
-        } else {
-          int32_t frame_time;
-          frame_time = max(1, 1000 / turns_per_second);
-          game.delta_time = 1;
-          LbSleepUntil(fe_last_loop_time + frame_time);
-        }
-      }
-      fe_last_loop_time = LbTimerClock();
-
-      api_update_server();
-
-    } while (!finish_menu);
-
-    LbPaletteFade(0, 8, Lb_PALETTE_FADE_CLOSED);
-    LbScreenClear(0);
-    LbScreenSwap();
-    FrontendMenuState prev_state;
-    prev_state = frontend_menu_state;
-    frontend_set_state(FeSt_INITIAL);
-    if (exit_keeper)
-    {
-      player = get_my_player();
-      player->display_flags &= ~PlaF6_PlyrHasQuit;
-      return true;
-    }
-    reenter_video_mode();
-
-    display_loading_screen();
-
-    short flgmem;
-    switch (prev_state)
-    {
-    case FeSt_START_KPRLEVEL:
-          my_player_number = default_loc_player;
-          game.game_kind = GKind_LocalGame;
-          clear_flag(game.system_flags, GSF_NetworkActive);
-          player = get_my_player();
-          player->is_active = 1;
-          startup_network_game(&loop, true);
-          break;
-    case FeSt_START_MPLEVEL:
-          set_flag(game.system_flags, GSF_NetworkActive);
-          skip_high_score_screen = 1;
-          game.game_kind = GKind_MultiGame;
-          player = get_my_player();
-          player->is_active = 1;
-          startup_network_game(&loop, false);
-          break;
-    case FeSt_LOAD_GAME:
-          flgmem = game.save_game_slot;
-          clear_flag(game.system_flags, GSF_NetworkActive);
-          LbScreenClear(0);
-          LbScreenSwap();
-          if (!load_game(game.save_game_slot))
-          {
-              ERRORLOG("Loading game %d failed; quitting.",(int)game.save_game_slot);
-              quit_game = 1;
-          }
-          game.save_game_slot = flgmem;
-          break;
-    case FeSt_PACKET_DEMO:
-          game.mode_flags |= MFlg_IsDemoMode;
-          startup_saved_packet_game();
-          set_gui_visible(false);
-          clear_flag(game.operation_flags, GOF_ShowPanel);
-          break;
-    }
-
-    coroutine_add(&loop, &set_not_has_quit);
-    coroutine_process(&loop);
-    if (loop.error)
-    {
-        frontend_set_state(FeSt_INITIAL);
-        return false;
-    }
-    return true;
-}
-
-void game_loop(void)
-{
-#if (BFDEBUG_LEVEL > 0)
-    unsigned long playtime = 0;
-#endif
-    SYNCDBG(0,"Entering gameplay loop.");
-
-    while ( !exit_keeper )
-    {
-      update_mouse();
-      while (!wait_at_frontend())
-      {
-          if (exit_keeper)
-              break;
-      }
-      if ( exit_keeper )
-        break;
-
-      int32_t mspos_x_bak = lbDisplay.MMouseX;
-      int32_t mspos_y_bak = lbDisplay.MMouseY;
-
-      if (game.game_kind == GKind_LocalGame)
-      {
-        if (game.save_game_slot == -1)
-        {
-            if (is_feature_on(Ft_SkipHeartZoom) == false) {
-                for (int i = 0; i < PLAYERS_COUNT; i++) {
-                    struct PlayerInfo *player = get_player(i);
-                    if (player_exists(player) && ((player->allocflags & PlaF_CompCtrl) == 0)) {
-                        set_player_instance(player, PI_HeartZoom, 0);
-                    }
-                }
-            } else {
-                if (!game.packet_load_enable) {
-                    toggle_status_menu(1); // Required when skipping PI_HeartZoom
-                }
-            }
-        } else
-        {
-          game.save_game_slot = -1;
-        }
-      } else {
-          for (int i = 0; i < PLAYERS_COUNT; i++) {
-              struct PlayerInfo *player = get_player(i);
-              if (player_exists(player) && ((player->allocflags & PlaF_CompCtrl) == 0)) {
-                  set_player_instance(player, PI_HeartZoom, 0);
-              }
-          }
-      }
-
-      // Try to keep the mouse position unchanged when entering the level.
-      // The main considerations are:
-      // 1. SKIP_HEART_ZOOM: the mouse icon position will be reset to the top-left corner (0, 0), but the actual mouse position remains unchanged.
-      // 2. PI_HeartZoom: the mouse will be moved to the center of the screen.
-      LbMouseSetPosition(mspos_x_bak, mspos_y_bak);
-
-      unsigned long starttime;
-#if (BFDEBUG_LEVEL > 0)
-      unsigned long endtime;
-#endif
-      struct Dungeon *dungeon;
-      // get_my_dungeon() can't be used here because players are not initialized yet
-      dungeon = get_dungeon(my_player_number);
-      starttime = LbTimerClock();
-      dungeon->lvstats.start_time = starttime;
-      dungeon->lvstats.end_time = starttime;
-      if (!TimerNoReset)
-      {
-          if (is_feature_on(Ft_SkipHeartZoom))
-          {
-              timerstarttime = starttime;
-          }
-          else
-          {
-              TimerFreeze = true;
-          }
-          memset(&Timer, 0, sizeof(Timer));
-      }
-      LbScreenClear(0);
-      LbScreenSwap();
-      game.frame_skip = 0;
-      keeper_gameplay_loop();
-      set_pointer_graphic_none();
-      LbScreenClear(0);
-      LbScreenSwap();
-      stop_atmos_sounds();
-      stop_music(true);
-      stop_streamed_samples();
-      free_level_strings_data();
-      turn_off_all_menus();
-      delete_all_structures();
-      clear_mapwho();
-      // Reset sounds back to the fxdata baseline so the main menu (and any
-      // subsequent campaign/freeplay selection) hears unmodified defaults.
-      sound_reset_to_fxdata_baseline();
-#if (BFDEBUG_LEVEL > 0)
-      endtime = LbTimerClock();
-#endif
-      quit_game = 0;
-      if ((game.operation_flags & GOF_SingleLevel) != 0)
-          exit_keeper=true;
-#if (BFDEBUG_LEVEL > 0)
-      playtime += endtime-starttime;
-#endif
-      SYNCDBG(0,"Play time is %lu seconds",playtime>>10);
-      reset_eye_lenses();
-      close_packet_file();
-      game.packet_load_enable = false;
-      game.packet_save_enable = false;
-    } // end while
-
-    // Stop the movie recording if it's on
-    if ((game.system_flags & GSF_CaptureMovie) != 0) {
-        movie_record_stop();
-    }
-    ShutDownSDLAudio();
-    SYNCDBG(7,"Done");
-}
-
-short reset_game(void)
-{
-    SYNCDBG(6,"Starting");
-
-    LbMouseSuspend();
-    LbIKeyboardClose();
-    LbScreenReset(false);
-    LbDataFreeAllV2(game_load_files);
-    free_gui_strings_data();
-    free_level_strings_data();
-    FreeAudio();
-    return 1;
-}
 
 short process_command_line(unsigned short argc, char *argv[])
 {
@@ -4492,6 +3051,20 @@ const char* determine_log_filename(unsigned short argument_count, char *argument
         }
     }
     return log_file_name;
+}
+
+static short reset_game(void)
+{
+    SYNCDBG(6,"Starting");
+
+    LbMouseSuspend();
+    LbIKeyboardClose();
+    LbScreenReset(false);
+    LbDataFreeAllV2(game_load_files);
+    free_gui_strings_data();
+    free_level_strings_data();
+    FreeAudio();
+    return 1;
 }
 
 int LbBullfrogMain(unsigned short argc, char *argv[])
